Record native comparison and Windows packaging evidence
Record package smoke results and deferred Windows tests

fix: restore progressive hiding and bound favorite prewarming

chore: stop tracking local scratch files

fix: rebase pending rotations after saving

test: settle sort commands before viewer cleanup

test: preserve cancellation assertions and await grouping

fix: refuse symlink destinations for image exports

fix: rebuild duplicate visibility after sorting

PR fixes

+++ internal/imaging/raw.go
@@ -192,7 +192,7 @@ func sliceJPEG(data []byte, off, length uint32) []byte {
 		return nil
 	}
 
-	blob := data[uint64(off) : uint64(off)+uint64(length)]
+	blob := data[off : uint64(off)+uint64(length)]
 	if !isJPEG(blob) {
 		return nil
 	}
@@ -228,7 +228,7 @@ func concatStrips(data []byte, offs, lens []uint32) []byte {
 		if !haveRange(data, off, lens[i]) {
 			return nil
 		}
-		out = append(out, data[uint64(off):uint64(off)+uint64(lens[i])]...)
+		out = append(out, data[off:uint64(off)+uint64(lens[i])]...)
 	}
 	return out
 }
